+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project are documented here. The format is
 based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and
 this project adheres to [Semantic Versioning](https://semver.org/).
+## [unreleased]
+
+### Bug Fixes
+
+- *(release)* Install syft before running goreleaser
+
 ## [0.2.2] - 2026-05-30
 
 ### Miscellaneous Tasks