+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 All notable changes to this project are documented here. The format is
 based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and
 this project adheres to [Semantic Versioning](https://semver.org/).
-## [unreleased]
+## [0.2.3] - 2026-05-30
 
 ### Bug Fixes
 
-- *(release)* Install syft before running goreleaser
+- *(release)* Install syft before goreleaser ([#33](https://github.com/donaldgifford/claudelint/issues/33))
 
 ## [0.2.2] - 2026-05-30